backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.2"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.48"}
alembic = "^1.18.4"
asyncpg = "^0.31.0"
pydantic-settings = "^2.13.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.22"
stripe = "^14.4.1"
httpx = "^0.28.1"
python-dotenv = "^1.2.2"
aiofiles = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
